requirements.txt
ping3
netmiko
requests
pandas
cryptography
python_dotenv
